requirements.txt
torch
torchvision
torchaudio
transformers
opencv-python
numpy
pillow
librosa
fastapi
uvicorn
ffmpeg-python
python-multipart